Move renderCodeError to lib
